Agrego parametro opcional last al Graphic, muevo funcion para zoom a modulo aparte, y disminuyo umbral de zoom
