+++ src/components/viewpage/graphic/GraphicAndShare.tsx
@@ -3,7 +3,6 @@ import * as moment from "moment";
 import * as React from 'react';
 import { connect } from "react-redux";
 import { setDate } from "../../../actions/seriesActions";
-import { IDataPoint } from "../../../api/DataPoint";
 import { IDateRange } from "../../../api/DateSerie";
 import QueryParams from '../../../api/QueryParams';
 import { ISerie } from "../../../api/Serie";
@@ -16,6 +15,7 @@ import GraphContainer from "../../style/Graphic/GraphContainer";
 import { getQueryParams } from '../ViewPage';
 import Graphic, { IChartExtremeProps } from "./Graphic";
 import GraphicComplements from "./GraphicComplements";
+import { chartExtremes } from '../../../helpers/graphic/chartExtremes';
 
 
 const DEFAULT_CHART_TYPE: string = 'line';
@@ -201,23 +201,6 @@ function findSerieDate(series: ISerie[], timestamp: number): string {
     const serieData = firstSerieData.find((data) => data.date >= formattedMoment(localTimestamp(timestamp)));
     return serieData !== undefined ? serieData.date : '';
 }
-
-
-export function chartExtremes(series: ISerie[], dateRange: { start: string, end: string }): IChartExtremeProps {
-    if (series.length === 0) {return {min: 0, max: 0}}
-
-    const firstSerieData = series[0].data;
-    let minDataIndex = firstSerieData.findIndex((data: IDataPoint) => data.date >= dateRange.start);
-    let maxDataIndex = firstSerieData.findIndex((data: IDataPoint) => data.date >= dateRange.end);
-    if (minDataIndex <= 0) { minDataIndex = 0 }
-    if (maxDataIndex <= 0) { maxDataIndex = firstSerieData.length - 1}
-
-    const min = new Date(firstSerieData[minDataIndex].date).getTime();
-    const max = new Date(firstSerieData[maxDataIndex].date).getTime();
-
-    return {min, max};
-}
-
 
 function mapStateToProps(state: IStore) {
     return {
